+++ Sources/Sentry/SentryANRTrackingIntegration.m
@@ -29,6 +29,7 @@ SentryANRTrackingIntegration ()
 
 @property (nonatomic, strong) SentryANRTracker *tracker;
 @property (nonatomic, strong) SentryOptions *options;
+@property (atomic, assign) BOOL reportAppHangs;
 
 @end
 
@@ -45,6 +46,7 @@ SentryANRTrackingIntegration ()
 
     [self.tracker addListener:self];
     self.options = options;
+    self.reportAppHangs = YES;
 
     return YES;
 }
@@ -52,6 +54,16 @@ SentryANRTrackingIntegration ()
 - (SentryIntegrationOption)integrationOptions
 {
     return kIntegrationOptionEnableAppHangTracking | kIntegrationOptionDebuggerNotAttached;
+}
+
+- (void)pauseAppHangTracking
+{
+    self.reportAppHangs = NO;
+}
+
+- (void)resumeAppHangTracking
+{
+    self.reportAppHangs = YES;
 }
 
 - (void)uninstall
@@ -66,6 +78,11 @@ SentryANRTrackingIntegration ()
 
 - (void)anrDetected
 {
+    if (self.reportAppHangs == NO) {
+        SENTRY_LOG_DEBUG(@"AppHangTracking paused. Ignoring reported app hang.")
+        return;
+    }
+
 #if SENTRY_HAS_UIKIT
     // If the app is not active, the main thread may be blocked or too busy.
     // Since there is no UI for the user to interact, there is no need to report app hang.